dev: removed support for configuring GIT_PATH and storing grs history.
Storing a history of grs runs was a motivating idea, but turned out to be not that interesting
Now that I've decided to stop doing development on Windows (cygwin), support for configurable GIT_PATH is useless.

Removing support for these two things, in addition to previous work to migrate to cobra+viper, allowed me to delete
a bunch of custom code. Hurry for deleting code!